+++ DataAccess/Models/Posting.cs
@@ -7,11 +7,11 @@ namespace DataAccess.Models
         public int PostingId { get; set; }
         public int UserId { get; set; }
         public int? SubcategoryId { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public DateTime Date { get; set; }
-        public double? Latitude { get; set; }
-        public double? Longitude { get; set; }
-        public double? Accuracy { get; set; }
+        public decimal? Latitude { get; set; }
+        public decimal? Longitude { get; set; }
+        public decimal? Accuracy { get; set; }
         public DateTime CreatedOn { get; set; }
         public DateTime UpdatedOn { get; set; }
     }
